+++ src/MultiphotonBreitWheeler/MultiphotonBreitWheeler.cpp
@@ -728,7 +728,9 @@ void MultiphotonBreitWheeler::removeDecayedPhotonsWithoutBinCompression(
                         Epart[iDim*nparts+ipart] = Epart[iDim*nparts+last_photon_index];
                         Bpart[iDim*nparts+ipart] = Bpart[iDim*nparts+last_photon_index];
                     }
-                    for ( int iDim=n_dimensions_-1 ; iDim>=0 ; iDim-- ) {
+                    int iDim_max_iold = n_dimensions_-1;
+                    if (thetaold) iDim_max_iold = 1; // in AMcylindrical iold and deltaold are defined in 2D
+                    for ( int iDim=iDim_max_iold ; iDim>=0 ; iDim-- ) {
                         iold[iDim*nparts+ipart] = iold[iDim*nparts+last_photon_index];
                         deltaold[iDim*nparts+ipart] = deltaold[iDim*nparts+last_photon_index];
                     }